+++ sdk/messaging/azservicebus/messageSettler_test.go
@@ -21,42 +21,41 @@ func TestMessageSettlementUsingReceiver(t *testing.T) {
 
 	err := testStuff.Sender.SendMessage(context.Background(), &Message{
 		Body: []byte("hello"),
-	})
+	}, nil)
 	require.NoError(t, err)
 
-	var msg *ReceivedMessage
-	msg, err = receiver.receiveMessage(ctx, nil)
+	messages, err := receiver.ReceiveMessages(ctx, 1, nil)
 	require.NoError(t, err)
-	require.EqualValues(t, 1, msg.DeliveryCount)
+	require.EqualValues(t, 1, messages[0].DeliveryCount)
 
 	// message from queue -> Abandon -> back to the queue
-	err = receiver.AbandonMessage(context.Background(), msg, &AbandonMessageOptions{
+	err = receiver.AbandonMessage(context.Background(), messages[0], &AbandonMessageOptions{
 		PropertiesToModify: map[string]interface{}{
 			"hello": "world",
 		},
 	})
 	require.NoError(t, err)
 
-	msg, err = receiver.receiveMessage(ctx, nil)
+	messages, err = receiver.ReceiveMessages(ctx, 1, nil)
 	require.NoError(t, err)
-	require.EqualValues(t, 2, msg.DeliveryCount)
-	require.EqualValues(t, "world", msg.ApplicationProperties["hello"].(string))
+	require.EqualValues(t, 2, messages[0].DeliveryCount)
+	require.EqualValues(t, "world", messages[0].ApplicationProperties["hello"].(string))
 
 	// message from queue -> DeadLetter -> to the dead letter queue
-	err = receiver.DeadLetterMessage(ctx, msg, &DeadLetterOptions{
+	err = receiver.DeadLetterMessage(ctx, messages[0], &DeadLetterOptions{
 		ErrorDescription: to.Ptr("the error description"),
 		Reason:           to.Ptr("the error reason"),
 	})
 	require.NoError(t, err)
 
-	msg, err = deadLetterReceiver.receiveMessage(ctx, nil)
+	messages, err = deadLetterReceiver.ReceiveMessages(ctx, 1, nil)
 	require.NoError(t, err)
-	require.EqualValues(t, 2, msg.DeliveryCount)
+	require.EqualValues(t, 2, messages[0].DeliveryCount)
 
-	require.EqualValues(t, "the error description", *msg.DeadLetterErrorDescription)
-	require.EqualValues(t, "the error reason", *msg.DeadLetterReason)
+	require.EqualValues(t, "the error description", *messages[0].DeadLetterErrorDescription)
+	require.EqualValues(t, "the error reason", *messages[0].DeadLetterReason)
 
-	require.EqualValues(t, *msg.ExpiresAt, msg.EnqueuedTime.Add(*msg.TimeToLive))
+	require.EqualValues(t, *messages[0].ExpiresAt, messages[0].EnqueuedTime.Add(*messages[0].TimeToLive))
 
 	// TODO: introducing deferred messages into the chain seems to have broken something.
 	// // message from dead letter queue -> Defer -> to the dead letter queue's deferred messages
@@ -67,20 +66,20 @@ func TestMessageSettlementUsingReceiver(t *testing.T) {
 	// require.NoError(t, err)
 
 	// deferred message from dead letter queue -> Abandon -> dead letter queue
-	err = deadLetterReceiver.AbandonMessage(ctx, msg, &AbandonMessageOptions{
+	err = deadLetterReceiver.AbandonMessage(ctx, messages[0], &AbandonMessageOptions{
 		PropertiesToModify: map[string]interface{}{
 			"hello": "world",
 		},
 	})
 	require.NoError(t, err)
 
-	msg, err = deadLetterReceiver.receiveMessage(ctx, nil)
+	messages, err = deadLetterReceiver.ReceiveMessages(ctx, 1, nil)
 	require.NoError(t, err)
-	require.EqualValues(t, 2, msg.DeliveryCount)
-	require.EqualValues(t, "world", msg.ApplicationProperties["hello"].(string))
+	require.EqualValues(t, 2, messages[0].DeliveryCount)
+	require.EqualValues(t, "world", messages[0].ApplicationProperties["hello"].(string))
 
 	// message from dead letter queue -> Complete -> (deleted from queue)
-	err = deadLetterReceiver.CompleteMessage(ctx, msg)
+	err = deadLetterReceiver.CompleteMessage(ctx, messages[0], nil)
 	require.NoError(t, err)
 }
 
@@ -98,7 +97,7 @@ func TestMessageSettlementUsingReceiverWithReceiveAndDelete(t *testing.T) {
 
 	err = sender.SendMessage(context.Background(), &Message{
 		Body: []byte("hello"),
-	})
+	}, nil)
 	require.NoError(t, err)
 
 	receiver, err := client.NewReceiverForQueue(queueName, &ReceiverOptions{
@@ -111,7 +110,7 @@ func TestMessageSettlementUsingReceiverWithReceiveAndDelete(t *testing.T) {
 	require.NotEmpty(t, messages)
 
 	require.EqualValues(t, internal.RecoveryKindFatal, internal.GetSBErrInfo(receiver.AbandonMessage(ctx, messages[0], nil)).RecoveryKind)
-	require.EqualValues(t, internal.RecoveryKindFatal, internal.GetSBErrInfo(receiver.CompleteMessage(ctx, messages[0])).RecoveryKind)
+	require.EqualValues(t, internal.RecoveryKindFatal, internal.GetSBErrInfo(receiver.CompleteMessage(ctx, messages[0], nil)).RecoveryKind)
 	require.EqualValues(t, internal.RecoveryKindFatal, internal.GetSBErrInfo(receiver.DeferMessage(ctx, messages[0], nil)).RecoveryKind)
 	require.EqualValues(t, internal.RecoveryKindFatal, internal.GetSBErrInfo(receiver.DeadLetterMessage(ctx, messages[0], nil)).RecoveryKind)
 
@@ -143,16 +142,15 @@ func TestDeferredMessages(t *testing.T) {
 		// BUG: we're timing out here, even though our abandon should have put the message
 		// back into the queue. It appears that settlement methods don't work on messages
 		// that have been received as deferred.
-		msg, err = receiver.receiveMessage(ctx, nil)
+		messages, err := receiver.ReceiveMessages(ctx, 1, nil)
 		require.NoError(t, err)
-		require.NotNil(t, msg)
-		require.EqualValues(t, "world", msg.ApplicationProperties["hello"].(string))
+		require.EqualValues(t, "world", messages[0].ApplicationProperties["hello"].(string))
 	})
 
 	t.Run("Complete", func(t *testing.T) {
 		msg := testStuff.deferMessageForTest(t)
 
-		err := receiver.CompleteMessage(ctx, msg)
+		err := receiver.CompleteMessage(ctx, msg, nil)
 		require.NoError(t, err)
 
 		assertEntityEmpty(t, receiver)
@@ -173,12 +171,12 @@ func TestDeferredMessages(t *testing.T) {
 		})
 		require.NoError(t, err)
 
-		msg, err = receiver.receiveDeferredMessage(ctx, *msg.SequenceNumber)
+		deferredMessages, err := receiver.ReceiveDeferredMessages(ctx, []int64{*msg.SequenceNumber}, nil)
 		require.NoError(t, err)
 
-		require.EqualValues(t, "world", msg.ApplicationProperties["hello"].(string))
+		require.EqualValues(t, "world", deferredMessages[0].ApplicationProperties["hello"].(string))
 
-		err = receiver.CompleteMessage(ctx, msg)
+		err = receiver.CompleteMessage(ctx, deferredMessages[0], nil)
 		require.NoError(t, err)
 
 		assertEntityEmpty(t, receiver)
@@ -200,12 +198,11 @@ func TestDeferredMessage_DeadLettering(t *testing.T) {
 	require.NoError(t, err)
 
 	// check that the message made it to the dead letter queue
-	msg, err = deadLetterReceiver.receiveMessage(context.Background(), nil)
+	messages, err := deadLetterReceiver.ReceiveMessages(context.Background(), 1, nil)
 	require.NoError(t, err)
-	require.NotNil(t, msg)
 
 	// remove it from the DLQ
-	require.NoError(t, deadLetterReceiver.CompleteMessage(context.Background(), msg))
+	require.NoError(t, deadLetterReceiver.CompleteMessage(context.Background(), messages[0], nil))
 
 	// and...everything should be clean
 	assertEntityEmpty(t, deadLetterReceiver)
@@ -226,33 +223,32 @@ func TestMessageSettlementUsingOnlyBackupSettlement(t *testing.T) {
 
 	err := testStuff.Sender.SendMessage(context.Background(), &Message{
 		Body: []byte("hello"),
-	})
+	}, nil)
 	require.NoError(t, err)
 
 	// toggle the super secret switch
 	actualSettler, _ = receiver.settler.(*messageSettler)
 	actualSettler.onlyDoBackupSettlement = true
 
-	var msg *ReceivedMessage
-	msg, err = receiver.receiveMessage(ctx, nil)
+	messages, err := receiver.ReceiveMessages(ctx, 1, nil)
 	require.NoError(t, err)
-	require.EqualValues(t, 1, msg.DeliveryCount)
+	require.EqualValues(t, 1, messages[0].DeliveryCount)
 
-	err = receiver.AbandonMessage(context.Background(), msg, nil)
-	require.NoError(t, err)
-
-	msg, err = receiver.receiveMessage(ctx, nil)
-	require.NoError(t, err)
-	require.EqualValues(t, 2, msg.DeliveryCount)
-
-	err = receiver.DeadLetterMessage(ctx, msg, nil)
+	err = receiver.AbandonMessage(context.Background(), messages[0], nil)
 	require.NoError(t, err)
 
-	msg, err = deadLetterReceiver.receiveMessage(ctx, nil)
+	messages, err = receiver.ReceiveMessages(ctx, 1, nil)
 	require.NoError(t, err)
-	require.EqualValues(t, 2, msg.DeliveryCount)
+	require.EqualValues(t, 2, messages[0].DeliveryCount)
 
-	err = deadLetterReceiver.CompleteMessage(context.Background(), msg)
+	err = receiver.DeadLetterMessage(ctx, messages[0], nil)
+	require.NoError(t, err)
+
+	messages, err = deadLetterReceiver.ReceiveMessages(ctx, 1, nil)
+	require.NoError(t, err)
+	require.EqualValues(t, 2, messages[0].DeliveryCount)
+
+	err = deadLetterReceiver.CompleteMessage(context.Background(), messages[0], nil)
 	require.NoError(t, err)
 }
 
@@ -315,21 +311,20 @@ func assertEntityEmpty(t *testing.T, receiver *Receiver) {
 func (testStuff *testStuff) deferMessageForTest(t *testing.T) *ReceivedMessage {
 	err := testStuff.Sender.SendMessage(context.Background(), &Message{
 		Body: []byte("hello"),
-	})
+	}, nil)
 	require.NoError(t, err)
 
-	var msg *ReceivedMessage
-	msg, err = testStuff.Receiver.receiveMessage(context.Background(), nil)
+	messages, err := testStuff.Receiver.ReceiveMessages(context.Background(), 1, nil)
 	require.NoError(t, err)
-	require.NotNil(t, msg)
+	require.NotEmpty(t, messages)
 
-	require.EqualValues(t, 1, msg.DeliveryCount)
+	require.EqualValues(t, 1, messages[0].DeliveryCount)
 
-	err = testStuff.Receiver.DeferMessage(context.Background(), msg, nil)
-	require.NoError(t, err)
-
-	msg, err = testStuff.Receiver.receiveDeferredMessage(context.Background(), *msg.SequenceNumber)
+	err = testStuff.Receiver.DeferMessage(context.Background(), messages[0], nil)
 	require.NoError(t, err)
 
-	return msg
+	messages, err = testStuff.Receiver.ReceiveDeferredMessages(context.Background(), []int64{*messages[0].SequenceNumber}, nil)
+	require.NoError(t, err)
+
+	return messages[0]
 }